Changing structure of files and fixing namespaces to match (going for a more PSR feel); Also fixed missing AlertNotAvailable exception, adjusted it to be 900 code and Information is now 1000
