--- config.php.sample
@@ -1,12 +1,0 @@
-<?php
-// Sample config.php 
-// This holds all variables that you may not want to get committed into a git repo. 
-// 
-// Rename this to config.php
-
-namespace libMBTA;
-
-// This is the testing key. DO NOT USE IN PRODUCTION.
-$APIKEY = "wX9NwuHnZU2ToO7GmGR9uw";
-
-?>
