+++ standaloneSA/libMBTA/config.php.sample
@@ -1,0 +1,11 @@
+<?php
+
+	// Sample standaloneSA\libMBTA\config.php
+	// This holds all variables that you may not want to get committed into a git repo. 
+	// 
+	// Rename this to config.php
+
+	namespace standaloneSA\libMBTA\libMBTA;
+
+	// This is the testing key. DO NOT USE IN PRODUCTION.
+	$APIKEY = "wX9NwuHnZU2ToO7GmGR9uw";
